feat(controller-refactor). Migrate guts of controller to separate module.
